+++ l10n/sv.js
@@ -53,7 +53,6 @@ OC.L10N.register(
     "Display the header more compact" : "Visa kompakt sidhuvud",
     "Display Feedback & Support menu button" : "Visa Feedback & Support i menyn",
     "Display Help menu button" : "Visa hjälp i menyn",
-    "Display toolbar tabs" : "Visa flikar i verktygsfältet",
     "Save as" : "Spara som",
     "File saved" : "Filen har sparats",
     "Insert image" : "Infoga bild",
@@ -79,6 +78,7 @@ OC.L10N.register(
     "Show watermark on link shares with specific system tags" : "Visa vattenmärke för länkdelningar med specifika systemtaggar",
     "Select tag" : "Välj tagg",
     "version": "verzia",
-    "Disable certificate verification (insecure)": "Använd inte certifikatsverifiering (osäkert)"
+    "Disable certificate verification (insecure)": "Använd inte certifikatsverifiering (osäkert)",
+    "Generate document preview": "Skapa förhandsgranskning av dokument"
 },
 "nplurals=2; plural=(n != 1);");